Evim, the vim shortcut I've always wanted, can list open vim instances, and open a file in one of those instances, if one doesn't exist, then it can prompt you to start one of your choice.
To keep things simple, I have elected to not cater for the desire to start
another vim instance once others are already open. I've only included the
start if one does not exist option in the offchance that someone might grow
accustomed to this shortcut and it would be convenient to start your first
session from here if you haven't done so already.

Over and out